+++ src/compas_fab/backends/ros/backend_features/move_it_inverse_kinematics.py
@@ -5,8 +5,8 @@ from __future__ import print_function
 from compas.utilities import await_callback
 
 from compas_fab.backends.interfaces import InverseKinematics
-from compas_fab.backends.ros.backend_features.helpers import validate_response
 from compas_fab.backends.ros.backend_features.helpers import convert_constraints_to_rosmsg
+from compas_fab.backends.ros.backend_features.helpers import validate_response
 from compas_fab.backends.ros.messages import AttachedCollisionObject
 from compas_fab.backends.ros.messages import GetPositionIKRequest
 from compas_fab.backends.ros.messages import GetPositionIKResponse
@@ -18,6 +18,7 @@ from compas_fab.backends.ros.messages import PoseStamped
 from compas_fab.backends.ros.messages import PositionIKRequest
 from compas_fab.backends.ros.messages import RobotState
 from compas_fab.backends.ros.service_description import ServiceDescription
+from compas_fab.robots import Duration
 
 __all__ = [
     'MoveItInverseKinematics',
@@ -61,10 +62,12 @@ class MoveItInverseKinematics(InverseKinematics):
             - ``"constraints"``: (:obj:`list` of :class:`compas_fab.robots.Constraint`, optional)
               A set of constraints that the request must obey.
               Defaults to ``None``.
-            - ``"attempts"``: (:obj:`int`, optional) The maximum number of inverse kinematic attempts.
-              Defaults to ``8``.
+            - ``"timeout"``: (:obj:`int`, optional) Maximum allowed time for inverse kinematic calculation in seconds.
+              Defaults to ``1``. This value supersedes the ``"attempts"`` argument used before ROS Noetic.
             - ``"attached_collision_meshes"``: (:obj:`list` of :class:`compas_fab.robots.AttachedCollisionMesh`, optional)
               Defaults to ``None``.
+            - ``"attempts"``: (:obj:`int`, optional) The maximum number of inverse kinematic attempts.
+              Defaults to ``8``. This value is ignored on ROS Noetic and newer, use ``"timeout"`` instead.
 
         Raises
         ------
@@ -108,12 +111,22 @@ class MoveItInverseKinematics(InverseKinematics):
 
         constraints = convert_constraints_to_rosmsg(options.get('constraints'), header)
 
+        timeout_in_secs = options.get('timeout', 1)
+        timeout_duration = Duration(timeout_in_secs, 0).to_data()
+
         ik_request = PositionIKRequest(group_name=group,
                                        robot_state=start_state,
                                        constraints=constraints,
                                        pose_stamped=pose_stamped,
                                        avoid_collisions=options.get('avoid_collisions', True),
-                                       attempts=options.get('attempts', 8))
+                                       attempts=options.get('attempts', 8),
+                                       timeout=timeout_duration)
+
+        # The field `attempts` was removed in Noetic (and higher)
+        # so it needs to be removed from the message otherwise it causes a serialization error
+        # https://github.com/ros-planning/moveit/pull/1288
+        if self.ros_client.ros_distro not in ('kinetic', 'melodic'):
+            del ik_request.attempts
 
         def convert_to_positions(response):
             callback((response.solution.joint_state.position, response.solution.joint_state.name))